webapi/tct-video-html5-tests/video/video_canPlayType_codecs.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
         Cui, Jieqiong <jieqiongx.cui@intel.com>

-->

<html>
  <head>
    <title>Video Test: video_canPlayType_codecs</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-html5-20121025/the-video-element.html#the-video-element" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if the video.canPlayType supports video/mp4 with different parameter" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <video id="m">
      Your browser does not support media elements.
    </video>
    <div id="log"></div>
    <script>
        var v = document.getElementById("m");
        test(function() {
            assert_true(v.canPlayType('video/mp4; codecs="avc1.42E01E, mp4a.40.2"') !== "", "MPEG-4 supported (codecs='avc1.42E01E, mp4a.40.2')");
        }, "Check if the video.canPlayType supports video/mp4 of codecs's parameter(avc1.42E01E, mp4a.40.2)");

        test(function() {
            assert_true(v.canPlayType('video/mp4; codecs="avc1.42E01E"') !== "", 'H.264 supported (codecs="avc1.42E01E")');
        }, "Check if the video.canPlayType supports video/mp4 of codecs's parameter(avc1.42E01E)");

        test(function() {
            assert_true(v.canPlayType('video/mp4; codecs="avc1.42E01E, mp4a.40"') !== "", "MPEG-4 supported (codecs='avc1.42E01E, mp4a.40')");
        }, "Check if the video.canPlayType supports video/mp4 of codecs's parameter(avc1.42E01E, mp4a.40)");
    </script>
  </body>
</html>
